# src/libs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "CORNERSTONE API",
    "description": "GUAM CORNERSTONE API",
    "contact": {
      "name": "API Support",
      "email": "support@guamcornerstone.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Bearer {token}"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get Current User",
        "description": "Get Current User",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Update Current User",
        "description": "Update Current User",
        "operationId": "updateCurrentUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "helloupdate@gmail.com"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "hello"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "update"
                  }
                }
              },
              "examples": {
                "Update Current User": {
                  "value": {
                    "email": "helloupdate@gmail.com",
                    "firstName": "hello",
                    "lastName": "update"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "description": "Login",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "admin@dev.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "123qwe"
                  }
                }
              },
              "examples": {
                "Login": {
                  "value": {
                    "email": "admin@dev.com",
                    "password": "123qwe"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register",
        "description": "Register",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "hello@gmail.com"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "hello"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "world"
                  },
                  "password": {
                    "type": "string",
                    "example": "123qwe"
                  }
                }
              },
              "examples": {
                "Register": {
                  "value": {
                    "email": "hello@gmail.com",
                    "firstName": "hello",
                    "lastName": "world",
                    "password": "123qwe"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/auth/me/password": {
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Change Password Current User",
        "description": "Change Password Current User",
        "operationId": "changePasswordCurrentUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "current": {
                    "type": "string",
                    "example": "123qwe"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456"
                  }
                }
              },
              "examples": {
                "Change Password Current User": {
                  "value": {
                    "current": "123qwe",
                    "password": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/roles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "List Roles",
        "description": "List Roles",
        "operationId": "listRoles",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "10"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "Create Role",
        "description": "Create Role",
        "operationId": "createRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "test"
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "USER"
                    },
                    "example": [
                      "USER"
                    ]
                  }
                }
              },
              "examples": {
                "Create Role": {
                  "value": {
                    "name": "test",
                    "permissions": [
                      "USER"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/roles/all": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Get All Roles",
        "description": "Get All Roles",
        "operationId": "allRoles",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/roles/permissions": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Get All Permissions",
        "description": "Get All Permissions",
        "operationId": "allPermissions",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/roles/{id}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "summary": "Get Role",
        "description": "Get Role",
        "operationId": "getRole",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "summary": "Update Role",
        "description": "Update Role",
        "operationId": "updateRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "test"
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "LAND"
                    },
                    "example": [
                      "LAND"
                    ]
                  }
                }
              },
              "examples": {
                "Update Role": {
                  "value": {
                    "name": "test",
                    "permissions": [
                      "LAND"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "summary": "Delete Role",
        "description": "Delete Role",
        "operationId": "deleteRole",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List Users",
        "description": "List Users",
        "operationId": "listUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create User",
        "description": "Create User",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "usertest@gmail.com"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "user"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "test"
                  },
                  "password": {
                    "type": "string",
                    "example": "123qwe"
                  },
                  "roleId": {
                    "type": "number",
                    "example": 13
                  }
                }
              },
              "examples": {
                "Create User": {
                  "value": {
                    "email": "usertest@gmail.com",
                    "firstName": "user",
                    "lastName": "test",
                    "password": "123qwe",
                    "roleId": 13
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "summary": "Get User",
        "description": "Get User",
        "operationId": "getUser",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "summary": "Update User",
        "description": "Update User",
        "operationId": "updateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "helloupdate@gmail.com"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "hello"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "update2"
                  },
                  "roleId": {
                    "type": "number",
                    "example": 11
                  }
                }
              },
              "examples": {
                "Update User": {
                  "value": {
                    "email": "helloupdate@gmail.com",
                    "firstName": "hello",
                    "lastName": "update2",
                    "roleId": 11
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "summary": "Delete User",
        "description": "Delete User",
        "operationId": "deleteUser",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
  "tags": []
}